server/requirements.txt

fastapi
uvicorn
sqlalchemy
alembic

python-jose
passlib
bcrypt
python-multipart
slowapi
python-dotenv
